requirements.txt
requests==2.31.0
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
networkx==3.2.1
scikit-learn==1.3.2
streamlit==1.28.2
plotly==5.18.0
tqdm==4.66.1
loguru==0.7.2
tenacity==8.2.3
